move Activity stuff into an Activity directory
