+++ src/components/shared/Activity/ActivityItem/TrustOperationActivity.tsx
@@ -7,9 +7,9 @@ import { TrustOperation } from "@raha/api-shared/dist/models/Operation";
 
 import { ActivityTemplate, ActivityTemplateView } from "./ActivityTemplate";
 import { MapStateToProps, connect } from "react-redux";
-import { RahaState } from "../../../store";
-import { Member } from "../../../store/reducers/members";
-import { getMemberById } from "../../../store/selectors/members";
+import { RahaState } from "../../../../store";
+import { Member } from "../../../../store/reducers/members";
+import { getMemberById } from "../../../../store/selectors/members";
 
 type OwnProps = {
   operation: TrustOperation;
