+++ src/components/shared/Navigation.tsx
@@ -25,13 +25,13 @@ import { Mint } from "../pages/Mint";
 import { LogIn } from "../pages/LogIn";
 import { PendingInvites } from "../pages/PendingInvites";
 import { Profile as ProfileScreen } from "../pages/Profile";
-import { getMemberById } from "../../../src/store/selectors/members";
-import { RahaState } from "../../../src/store";
+import { getMemberById } from "../../store/selectors/members";
+import { RahaState } from "../../store";
 import { MemberList as MemberListScreen } from "../pages/MemberList";
 import { Onboarding } from "../pages/Onboarding/Onboarding";
 import { ReferralBonus } from "../pages/ReferralBonus";
 import { getLoggedInFirebaseUserId } from "../../store/selectors/authentication";
-import { Button, Text } from "../shared/elements";
+import { Button, Text } from "./elements";
 import { Discover } from "../pages/Discover";
 import { LeaderBoard } from "../pages/LeaderBoard";
 import { Invite } from "../pages/Invite/Invite";
